+++ newsletter-bundle/CHANGELOG.md
@@ -2,5 +2,6 @@
 
 ## DEV
 
+ * Remove subscriptions that are not activated within 24 hours (see contao/core-bundle#1512).
  * Add a custom text field to the subscribe module (see contao/core-bundle#1512).
  * Use type="email" in the default template (see #20).